Add files via upload
This is the first of my project on github, made it a year ago. This contains a russian roulette computer game with very basic 11th grade coding logic. The game is not so interactive but works for logic. Please help me find any errors and understand more about logics, also help me with if you have simple or efficient logic.